拆分 leanengine.js 中的中间件到 middleware 目录下单独的文件；重构初始化方式。
