+++ lib/leanengine.js
@@ -2,14 +2,13 @@
 var connect = require('connect'),
   bodyParser = require('body-parser'),
   https = require('https'),
-  domain = require('domain'),
-  crypto = require('crypto'),
   timeout = require('connect-timeout'),
-  version = require('../package.json').version,
   AV = require('./av-extra'),
   avosExpressCookieSession = require('../middleware/cookie-session'),
   avosExpressHttpsRedirect = require('../middleware/https-redirect'),
-  util = require('util');
+  utils = require('./utils');
+
+var _ = AV._;
 
 if (process.env.LC_API_SERVER) {
   AV.serverURL = process.env.LC_API_SERVER;
@@ -17,260 +16,122 @@ if (process.env.LC_API_SERVER) {
 var NODE_ENV = process.env.NODE_ENV || 'development';
 AV.Cloud.__prod = NODE_ENV === 'production' ? 1 : 0;
 
+AV.express = function(options) {
+  var router = connect();
+
+  router.use(require('../middleware/health-check')());
+
+  ['1', '1.1'].forEach(function(apiVersion) {
+    router.use('/' + apiVersion + '/call', function(req, res, next) {
+      req.rpcCall = true;
+      next();
+    });
+
+    ['functions', 'call'].forEach(function(urlEndpoint) {
+      router.use('/' + apiVersion + '/' + urlEndpoint, createCloudFunctionRouter(options));
+    });
+  });
+
+  return router;
+};
+
 // override AV.Cloud to a connect app
 if (!AV._old_Cloud) {
   AV._old_Cloud = AV.Cloud;
-  var Cloud = AV.Cloud = connect();
+  var Cloud = AV.Cloud = AV.express();
   for (var key in AV._old_Cloud) {
     Cloud[key] = AV._old_Cloud[key];
   }
 }
 
-AV._.extend(AV.Cloud, require('./cloud'));
+_.extend(AV.Cloud, require('./cloud'));
 
 // Don't reject unauthorized ssl.
 if (https.globalAgent && https.globalAgent.options) {
   https.globalAgent.options.rejectUnauthorized = false;
 }
 
-var TIMEOUT = '15s';
-
 AV.Cloud.CookieSession = avosExpressCookieSession(AV);
 AV.Cloud.HttpsRedirect = avosExpressHttpsRedirect(AV);
 
-// 健康监测 router
-Cloud.use('/__engine/1/ping', function(req, res) {
-  res.end(JSON.stringify({
-    "runtime": "nodejs-" + process.version,
-    "version": version
-  }));
-});
+function createCloudFunctionRouter(options) {
+  options = options || {};
 
-['1', '1.1'].forEach(function(apiVersion) {
-  ['functions', 'call'].forEach(function(urlEndpoint) {
-    var route = '/' + apiVersion + '/' + urlEndpoint;
+  var cloudFunctions = connect();
 
-    // CORS middleware
-    Cloud.use(route, function(req, res, next) {
-      res.setHeader('Access-Control-Allow-Origin', req.headers.origin || '*');
-      if (req.method.toLowerCase() === 'options') {
-        res.setHeader('Access-Control-Max-Age','86400');
-        res.setHeader('Access-Control-Allow-Methods', 'PUT, GET, POST, DELETE, OPTIONS');
-        res.setHeader('Access-Control-Allow-Headers', 'X-LC-Id, X-LC-Key, X-LC-Session, X-LC-Sign, X-LC-Prod, X-Uluru-Application-Key, X-Uluru-Application-Id, X-Uluru-Application-Production, X-Uluru-Client-Version, X-Uluru-Session-Token, X-AVOSCloud-Application-Key, X-AVOSCloud-Application-Id, X-AVOSCloud-Application-Production, X-AVOSCloud-Client-Version, X-AVOSCloud-Session-Token, X-AVOSCloud-Super-Key, X-Requested-With, Content-Type, X-AVOSCloud-Request-sign');
-        res.setHeader('Content-Length', 0);
-        return res.end();
-      }
-      next();
-    });
+  cloudFunctions.use(timeout(options.timeout || '15s'));
+  cloudFunctions.use(bodyParser.urlencoded({extended: false, limit: '20mb'}));
+  cloudFunctions.use(bodyParser.json({limit: '20mb'}));
+  cloudFunctions.use(bodyParser.text({limit: '20mb'}));
+  cloudFunctions.use(require('../middleware/cors')());
+  cloudFunctions.use(require('../middleware/domain-wrapper')());
+  cloudFunctions.use(require('../middleware/parse-leancloud-headers')(AV, {restrict: true}));
 
-    Cloud.use(route, timeout(TIMEOUT));
-    Cloud.use(route, bodyParser.urlencoded({extended: false, limit: '20mb'}));
-    Cloud.use(route, bodyParser.json({limit: '20mb'}));
-    Cloud.use(route, bodyParser.text({limit: '20mb'}));
-
-    // domainWrapper
-    Cloud.use(route, function(req, res, next) {
-      if (process.domain) {
-        return next();
-      }
-      var d = domain.create();
-      d.add(req);
-      d.add(res);
-      d.on('error', function(err) {
-        console.error('LeanEngine function uncaughtException url=%s, msg=%s', req.originalUrl, err.stack || err.message || err);
-        if(!res.finished) {
-          res.statusCode = 500;
-          res.setHeader('content-type', 'application/json; charset=UTF-8');
-          res.end(JSON.stringify({code: 1, error: 'LeanEngine function uncaughtException'}));
-        }
-      });
-      d.run(next);
-    });
-
-    // parse authInfo
-    Cloud.use(route, function(req, res, next) {
-      var appId, appKey, masterKey, contentType, param, prod, prodHeader, prodValue, sessionToken;
-      contentType = req.headers['content-type'];
-      if (/^text\/plain.*/i.test(contentType)) {
-        if (req.body && req.body !== '') {
-          req.body = JSON.parse(req.body);
-        }
-        appId = req.body._ApplicationId;
-        appKey = req.body._ApplicationKey;
-        masterKey = req.body._MasterKey;
-        prodValue = req.body._ApplicationProduction;
-        sessionToken = req.body._SessionToken;
-        for (param in req.body) {
-          // remove _* but keep __*
-          if (param.charAt(0) === '_' && param.charAt(1) !== '_') {
-            delete req.body[param];
-          }
-        }
-        prod = 1;
-        if (prodValue === 0 || prodValue === false) {
-          prod = 0;
-        }
-        req.AV = {
-          id: appId,
-          key: appKey,
-          masterKey: masterKey,
-          prod: prod,
-          sessionToken: sessionToken
-        };
-      } else {
-        appId = req.headers['x-lc-id'] ||
-          req.headers['x-avoscloud-application-id'] ||
-          req.headers['x-uluru-application-id'];
-        appKey = req.headers['x-lc-key'] ||
-          req.headers['x-avoscloud-application-key'] ||
-          req.headers['x-uluru-application-key'];
-        masterKey = req.headers['x-avoscloud-master-key'] || req.headers['x-uluru-master-key'];
-        prodHeader = req.headers['x-lc-prod'] ||
-          req.headers['x-avoscloud-application-production'] ||
-          req.headers['x-uluru-application-production'];
-        sessionToken = req.headers['x-lc-session'] ||
-          req.headers['x-uluru-session-token'] ||
-          req.headers['x-avoscloud-session-token'];
-        prod = 1;
-        if (prodHeader === '0' || prodHeader === 'false') {
-          prod = 0;
-        }
-        if (appKey && (appKey.indexOf(',master') > 0)) {
-          masterKey = appKey.slice(0, appKey.indexOf(','));
-          appKey = null;
-        }
-        req.AV = {
-          id: appId,
-          key: appKey,
-          masterKey: masterKey,
-          prod: prod,
-          sessionToken: sessionToken
-        };
-      }
-      return next();
-    });
-
-    // authorization
-    Cloud.use(route, function(req, res, next) {
-      var key, master, requestSign, sign, timestamp, validSign, _ref;
-      if (!req.AV.id) {
-        return unauthResp(res);
-      }
-      if (AV.applicationId === req.AV.id &&
-          (AV.applicationKey === req.AV.key ||
-           AV.masterKey === req.AV.key ||
-           AV.masterKey === req.AV.masterKey)) {
-        if (AV.masterKey === req.AV.masterKey) {
-          req.AV.authMasterKey = true;
-        }
-        return next();
-      }
-      requestSign = req.headers['x-lc-sign'] || req.headers['x-avoscloud-request-sign'];
-      if (requestSign) {
-        _ref = requestSign.split(',');
-        sign = _ref[0];
-        timestamp = _ref[1];
-        master = _ref[2];
-        key = master === 'master' ? AV.masterKey : AV.applicationKey;
-        validSign = signByKey(timestamp, key);
-        if (validSign === sign.toLowerCase()) {
-          if (master === 'master') {
-            req.AV.authMasterKey = true;
-            req.AV.masterKey = key;
-          } else {
-            req.AV.key = key;
-          }
-          return next();
-        }
-      }
-      return unauthResp(res);
-    });
-
-    // get metadatas func
-    Cloud.use(route + '/_ops/metadatas', function(req, res) {
-      if (req.AV.authMasterKey) {
-        return resp(res, Object.keys(Cloud.__code));
-      }
-      return unauthResp(res);
-    });
-
-    // parseUserInfo
-    Cloud.use(route, function(req, res, next) {
-      if (req.AV.sessionToken && req.AV.sessionToken !== '') {
-        AV.User.become(req.AV.sessionToken, {
-          success: function(user) {
-            req.AV.user = user;
-            next();
-          },
-          error: function(user, err) {
-            next(err);
-          }
-        });
-      } else if (req.body.user) {
-        var userObj = new AV.User();
-        userObj._finishFetch(req.body.user, true);
-        req.AV.user = userObj;
-        return next();
-      } else {
-        return next();
-      }
-    });
-
-    Cloud.use(route, function(req, res) {
-      if (req.url === '/') {
-        return respError(res, 'no function or class name');
-      }
-
-      var cb = function(err, data, isBare) {
-        if (err) {
-          return respError(res, err);
-        }
-        if (isBare) {
-          return respBare(res, data);
-        } else {
-          return resp(res, data);
-        }
-      };
-
-      var split = req.url.split('/');
-      if (split.length == 2) { // cloud function
-        call(split[1], req.body, req.AV.user, req, {
-          decodeAVObject: urlEndpoint == 'call'
-        }, function(err, data) {
-          cb(err, data);
-        });
-      } else if (split.length == 3) { // class hook
-        var userObj = new AV.User();
-        if (split[1] === 'onVerified') {
-          userObj._finishFetch(req.body.object, true);
-          onVerified(req, split[2], userObj);
-          cb(null, 'ok');
-
-        } else if (split[1] === '_User' && split[2] === 'onLogin') {
-          userObj._finishFetch(req.body.object, true);
-          onLogin(req, userObj, cb);
-
-        } else if ((split[1] === 'BigQuery' || split[1] === 'Insight' ) && split[2] === 'onComplete') {
-          onCompleteBigQueryJob(req.body);
-          cb(null, 'ok');
-
-        } else {
-          classHook(split[1], hookNameMapping[split[2]], req.body.object, req.AV.user, req, cb);
-        }
-      }
-    });
-
-    // next 参数即使不使用，也一定要存在，否则 error handler 不生效
-    Cloud.use(route, function(err, req, res, next) { // jshint ignore:line
-      if(req.timedout) {
-        console.error('LeanEngine function timeout, url=%s, timeout=%d', req.originalUrl, err.timeout);
-        err.code = 124; // https://leancloud.cn/docs/error_code.html#_124
-        err.message = 'The request timed out on the server.';
-      }
-      respError(res, err);
-    });
+  cloudFunctions.use('/_ops/metadatas', function(req, res) {
+    if (req.AV.authMasterKey) {
+      return resp(res, Object.keys(Cloud.__code));
+    } else {
+      return utils.unauthResp(res);
+    }
   });
-});
+
+  cloudFunctions.use(require('../middleware/fetch-user')(AV));
+
+  cloudFunctions.use(function(req, res) {
+    if (req.url === '/') {
+      return respError(res, 'no function or class name');
+    }
+
+    var cb = function(err, data, isBare) {
+      if (err) {
+        return respError(res, err);
+      }
+      if (isBare) {
+        return respBare(res, data);
+      } else {
+        return resp(res, data);
+      }
+    };
+
+    var split = req.url.split('/');
+    if (split.length == 2) { // cloud function
+      call(split[1], req.body, req.AV.user, req, {
+        decodeAVObject: req.rpcCall
+      }, function(err, data) {
+        cb(err, data);
+      });
+    } else if (split.length == 3) { // class hook
+      var userObj = new AV.User();
+      if (split[1] === 'onVerified') {
+        userObj._finishFetch(req.body.object, true);
+        onVerified(req, split[2], userObj);
+        cb(null, 'ok');
+
+      } else if (split[1] === '_User' && split[2] === 'onLogin') {
+        userObj._finishFetch(req.body.object, true);
+        onLogin(req, userObj, cb);
+
+      } else if ((split[1] === 'BigQuery' || split[1] === 'Insight' ) && split[2] === 'onComplete') {
+        onCompleteBigQueryJob(req.body);
+        cb(null, 'ok');
+
+      } else {
+        classHook(split[1], hookNameMapping[split[2]], req.body.object, req.AV.user, req, cb);
+      }
+    }
+  });
+
+  cloudFunctions.use(function(err, req, res, next) { // jshint ignore:line
+    if(req.timedout) {
+      console.error('LeanEngine function timeout, url=%s, timeout=%d', req.originalUrl, err.timeout);
+      err.code = 124; // https://leancloud.cn/docs/error_code.html#_124
+      err.message = 'The request timed out on the server.';
+    }
+    respError(res, err);
+  });
+
+  return cloudFunctions;
+}
 
 var resp = function(res, data) {
   res.setHeader('Content-Type', 'application/json; charset=UTF-8');
@@ -336,7 +197,7 @@ var classHook = function(className, hook, object, user, req, cb) {
     return cb(err);
   }
 
-  var obj = decodeParams(AV._.extend({}, object, {
+  var obj = decodeParams(_.extend({}, object, {
     __type: 'Object',
     className: className
   }));
@@ -446,16 +307,6 @@ var hookNameMapping = {
   afterDelete: '__after_delete_for_'
 };
 
-var unauthResp = function(res) {
-  res.statusCode = 401;
-  res.setHeader('content-type', 'application/json; charset=UTF-8');
-  return res.end(JSON.stringify({ code: 401, error: 'Unauthorized.' }));
-};
-
-var signByKey = function(timestamp, key) {
-  return crypto.createHash('md5').update('' + timestamp + key).digest('hex');
-};
-
 /* options: req, user, params, object*/
 var prepareRequestObject = function(options) {
   var req = options.req;
@@ -497,12 +348,12 @@ var encodeResult = function(result) {
       object = object._toFullJSON([]);
     }
 
-    return AV._.mapObject(object, function(value) {
+    return _.mapObject(object, function(value) {
       return AV._encode(value, []);
     });
   };
 
-  if (util.isArray(result)) {
+  if (_.isArray(result)) {
     return result.map(function(object) {
       return encodeAVObject(object);
     });
