+++ lib/utils.js
@@ -1,4 +1,5 @@
 'use strict';
+
 exports.typeOf = function(obj) {
   var classToType;
   if (obj === void 0 || obj === null) {
@@ -15,4 +16,10 @@ exports.typeOf = function(obj) {
     '[object Object]': 'object'
   };
   return classToType[Object.prototype.toString.call(obj)];
+};
+
+exports.unauthResp = function(res) {
+  res.statusCode = 401;
+  res.setHeader('content-type', 'application/json; charset=UTF-8');
+  return res.end(JSON.stringify({ code: 401, error: 'Unauthorized.' }));
 };
